Bug fixes and Implemented key mappings for main keys
Fixed a bug where the DDR setting wasn't being set on both
instances of Port A in the System VIA. Setting the DDR now sets Port A
to `| !(ddr_val)`, setting all the read bits to `1`. This is how
*jsbeeb* seems to do it and I've had some success after doing this.

Even though the emulator is now correctly scanning the keyboard
and recognising key presses, it's still not outputting the chars
to the screen. More investigation required.

I've also changed the debugger front end to keep the keypresses in its
buffer while the emulator is paused. Originally, they were being cleared
on each frame, meaning it was hard to capture keypresses.

Finally, in an attempt to force the OS to copy the keypresses to the
screen, I've just hard-coded a V-Sync into the System VIA.